+++ packages/components/src/Form/Fields/FieldRangeSlider/FieldRangeSlider.tsx
@@ -24,18 +24,36 @@
 
  */
 
-export * from './FieldCheckbox'
-export * from './FieldCheckboxGroup'
-export * from './FieldColor'
-export * from './FieldRadio'
-export * from './FieldRadioGroup'
-export * from './FieldRangeSlider'
-export * from './FieldSelect'
-export * from './FieldSelectMulti'
-export * from './FieldSlider'
-export * from './FieldText'
-export * from './FieldTime'
-export * from './FieldTimeSelect'
-export * from './FieldTextArea'
-export * from './FieldToggleSwitch'
-export * from './Field'
+import React, { forwardRef, Ref } from 'react'
+import styled from 'styled-components'
+import { useID } from '../../../utils'
+import { RangeSlider, RangeSliderProps } from '../../Inputs/RangeSlider'
+import { Field, FieldProps, omitFieldProps, pickFieldProps } from '../Field'
+
+export interface FieldRangeSliderProps
+  extends RangeSliderProps,
+    Omit<FieldProps, 'validationMessage'> {}
+
+const FieldRangeSliderComponent = forwardRef(
+  (props: FieldRangeSliderProps, ref: Ref<HTMLInputElement>) => {
+    const id = useID(props.id)
+    return (
+      <Field
+        data-testid="FieldRangeSliderId"
+        {...pickFieldProps(props)}
+        id={id}
+      >
+        <RangeSlider
+          {...omitFieldProps(props)}
+          aria-describedby={`${id}-describedby`}
+          id={id}
+          ref={ref}
+        />
+      </Field>
+    )
+  }
+)
+
+FieldRangeSliderComponent.displayName = 'FieldRangeSliderComponent'
+
+export const FieldRangeSlider = styled(FieldRangeSliderComponent)``