+++ packages/components/src/Form/Fields/FieldSlider/FieldSlider.tsx
@@ -24,18 +24,32 @@
 
  */
 
-export * from './FieldCheckbox'
-export * from './FieldCheckboxGroup'
-export * from './FieldColor'
-export * from './FieldRadio'
-export * from './FieldRadioGroup'
-export * from './FieldRangeSlider'
-export * from './FieldSelect'
-export * from './FieldSelectMulti'
-export * from './FieldSlider'
-export * from './FieldText'
-export * from './FieldTime'
-export * from './FieldTimeSelect'
-export * from './FieldTextArea'
-export * from './FieldToggleSwitch'
-export * from './Field'
+import React, { forwardRef, Ref } from 'react'
+import styled from 'styled-components'
+import { useID } from '../../../utils'
+import { Slider, SliderProps } from '../../Inputs/Slider'
+import { Field, FieldProps, omitFieldProps, pickFieldProps } from '../Field'
+
+export interface FieldSliderProps
+  extends SliderProps,
+    Omit<FieldProps, 'validationMessage'> {}
+
+const FieldSliderComponent = forwardRef(
+  (props: FieldSliderProps, ref: Ref<HTMLInputElement>) => {
+    const id = useID(props.id)
+    return (
+      <Field data-testid="FieldSliderId" {...pickFieldProps(props)} id={id}>
+        <Slider
+          {...omitFieldProps(props)}
+          aria-describedby={`${id}-describedby`}
+          id={id}
+          ref={ref}
+        />
+      </Field>
+    )
+  }
+)
+
+FieldSliderComponent.displayName = 'FieldSliderComponent'
+
+export const FieldSlider = styled(FieldSliderComponent)``